pyproject.toml
[tool.poetry]
name = "youtube_api"
version = "0.1.0"
description = ""
authors = ["Jatin Katyal <jatin.katyal13@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.4"
django-rest-framework = "^0.1.0"
python-dotenv = "^0.17.1"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]
black = "^21.6b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
